20w45a

+++ src/main/java/net/earthcomputer/multiconnect/protocols/v1_13_2/mixin/LightUpdatePacketAccessor.java
@@ -16,10 +16,10 @@ public interface LightUpdatePacketAccessor {
     void setChunkZ(int chunkZ);
 
     @Accessor
-    void setSkyLightMask(int skylightMask);
+    void setSkyLightMask(long skylightMask);
 
     @Accessor
-    void setBlockLightMask(int blocklightMask);
+    void setBlockLightMask(long blocklightMask);
 
     @Accessor
     void setBlockLightUpdates(List<byte[]> blockLightUpdates);
